# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnFlow API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/account/me": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_account_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/account/profile-image": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Update Profile Image",
        "operationId": "update_profile_image_api_v1_account_profile_image_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileImageUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/account/skater-profile": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Skater Profile",
        "description": "200 + JSON null before a row exists. 404 here would look like 'not deployed' to the client.",
        "operationId": "get_skater_profile_api_v1_account_skater_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SkaterProfileOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Skater Profile Api V1 Account Skater Profile Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "account"
        ],
        "summary": "Patch Skater Profile",
        "operationId": "patch_skater_profile_api_v1_account_skater_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkaterProfilePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkaterProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/account": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Delete My Account",
        "operationId": "delete_my_account_api_v1_account_delete",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/account/export": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Export My Data",
        "operationId": "export_my_data_api_v1_account_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Export My Data Api V1 Account Export Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/account/quota": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Account Quota",
        "operationId": "get_account_quota_api_v1_account_quota_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountQuotaResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountQuotaResponse": {
        "properties": {
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "analyses_remaining": {
            "type": "integer",
            "title": "Analyses Remaining"
          },
          "trial_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Expires At"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "bonus_analyses": {
            "type": "integer",
            "title": "Bonus Analyses"
          },
          "monthly_free_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Free Remaining",
            "description": "None when Pro/Coach (unlimited analyses)."
          }
        },
        "type": "object",
        "required": [
          "tier",
          "analyses_remaining",
          "subscription_status",
          "bonus_analyses"
        ],
        "title": "AccountQuotaResponse"
      },
...
      "ConsentStatus": {
        "properties": {
          "granted": {
            "type": "boolean",
            "title": "Granted"
          },
          "granted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Granted At"
          },
          "copy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Copy Version"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "granted",
          "granted_at",
          "copy_version",
          "source"
        ],
        "title": "ConsentStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "profile_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Image Url"
          },
          "consent": {
            "$ref": "#/components/schemas/ConsentStatus"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "tier",
          "consent"
        ],
        "title": "MeResponse"
      },
...
      "ProfileImageUpdateRequest": {
        "properties": {
          "profile_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Image Url"
          }
        },
        "type": "object",
        "title": "ProfileImageUpdateRequest"
      },
...
      "SkaterProfileOut": {
        "properties": {
          "natural_stance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Natural Stance"
          },
          "skate_styles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skate Styles"
          },
          "primary_skate_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Skate Style"
          },
          "experience_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "age_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Range"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "home_park": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Park"
          },
          "favorite_brands": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Favorite Brands"
          },
          "onboarding_completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Completed At"
          }
        },
        "type": "object",
        "title": "SkaterProfileOut"
      },
      "SkaterProfilePatchRequest": {
        "properties": {
          "natural_stance": {
            "title": "Natural Stance"
          },
          "skate_styles": {
            "title": "Skate Styles"
          },
          "primary_skate_style": {
            "title": "Primary Skate Style"
          },
          "experience_level": {
            "title": "Experience Level"
          },
          "age_range": {
            "title": "Age Range"
          },
          "city": {
            "title": "City"
          },
          "home_park": {
            "title": "Home Park"
          },
          "favorite_brands": {
            "title": "Favorite Brands"
          },
          "complete_onboarding": {
            "title": "Complete Onboarding",
            "default": false
          }
        },
        "type": "object",
        "title": "SkaterProfilePatchRequest",
        "description": "Loose body — enum/limit/completion rules live in ``app.services.skater_profile``."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}